# refactor(tables): improve `tables` performance