--- a/src/common/meta/src/key/flow/flow_route.rs
+++ b/src/common/meta/src/key/flow/flow_route.rs
@@ -180,7 +180,8 @@ impl FlowRouteManager {
             req,
             DEFAULT_PAGE_SIZE,
             Arc::new(flow_route_decoder),
-        );
+        )
+        .into_stream();
 
         Box::pin(stream)
     }