--- a/src/common/meta/src/key/flow/table_flow.rs
+++ b/src/common/meta/src/key/flow/table_flow.rs
@@ -207,7 +207,8 @@ impl TableFlowManager {
             req,
             DEFAULT_PAGE_SIZE,
             Arc::new(table_flow_decoder),
-        );
+        )
+        .into_stream();
 
         Box::pin(stream)
     }