--- a/src/common/meta/src/key/table_name.rs
+++ b/src/common/meta/src/key/table_name.rs
@@ -259,7 +259,8 @@ impl TableNameManager {
             req,
             DEFAULT_PAGE_SIZE,
             Arc::new(table_decoder),
-        );
+        )
+        .into_stream();
 
         Box::pin(stream)
     }