I hereby agree to the terms of the [GreptimeDB CLA](https://github.com/GreptimeTeam/.github/blob/main/CLA.md).

## Refer to a related PR or issue link (optional)

## What's changed and what's your intention?

This patch tries to solve some user's complaint about the lag while switching database in mysql cli by enabling concurrent tables iteration in information_schema to accelerate table rehash.

It also introduces some refactor regarding reading table information in kvbackend.

For a database with 100k tables, it cuts `show tables` cost time from 430ms to 200ms.
```
# From
$ time mysql -h 127.0.0.1 -P 4002 -e "show tables" > /dev/null 2>&1
mysql -h 127.0.0.1 -P 4002 -e "show tables" > /dev/null 2>&1  0.01s user 0.00s system 7% cpu 0.432 total

# To
$ time mysql -h 127.0.0.1 -P 4002 -e "show tables" > /dev/null 2>&1
mysql -h 127.0.0.1 -P 4002 -e "show tables" > /dev/null 2>&1  0.01s user 0.00s system 7% cpu 0.201 total
```
## Checklist

- [ ] I have written the necessary rustdoc comments.
- [ ] I have added the necessary unit tests and integration tests.
- [ ] This PR requires documentation updates.